Move ServerCapabilities to DTO directory
