+++ tests/Unit/Client/Interceptor/SystemInfoInterceptorTestCase.php
@@ -7,8 +7,8 @@ namespace Temporal\Tests\Unit\Client\Interceptor;
 use PHPUnit\Framework\TestCase;
 use Temporal\Api\Workflowservice\V1\GetSystemInfoResponse;
 use Temporal\Api\Workflowservice\V1\GetSystemInfoResponse\Capabilities;
+use Temporal\Client\DTO\ServerCapabilities;
 use Temporal\Client\GRPC\ContextInterface;
-use Temporal\Client\GRPC\ServerCapabilities;
 use Temporal\Client\GRPC\ServiceClient;
 use Temporal\Client\GRPC\StatusCode;
 use Temporal\Client\Interceptor\SystemInfoInterceptor;
